requirements.txt
asgiref==3.8.1
Django==5.1.7
django-ckeditor-5==0.2.17
pillow==11.1.0
python-dotenv==1.0.1
sqlparse==0.5.3
